optimize(vLLM): move PostModel out
