+++ ChatTTS/model/gpt.py
@@ -1,9 +1,8 @@
-import os, platform
+import platform
 from dataclasses import dataclass
 import logging
 from typing import Union, List, Optional, Tuple, Callable
 import gc
-from pathlib import Path
 
 import torch
 import torch.nn as nn
