+++ ChatTTS/model/velocity/__init__.py
@@ -1,3 +1,2 @@
 from .llm import LLM
-from .post_model import PostModel
 from .sampling_params import SamplingParams
